Require container gateway on content/container but not foreman
* Update container gateway tests to match
* Update features in yaml to be dictionaries
