+++ src/vars/flavors/foreman-proxy-content.yml
@@ -1,7 +1,6 @@
 flavor_features:
   - foreman-proxy
   - pulp
-  - container-gateway
   - content/rpm
   - content/deb
   - content/container
